#3633 – added mode layer
