+++ packages/ketcher-core/src/application/editor/modes/index.ts
@@ -1,0 +1,1 @@
+export * from './SnakeMode';
