+++ packages/ketcher-polymer-editor-react/src/state/common/editorSlice.ts
@@ -25,7 +25,6 @@ interface EditorState {
   activeTool: string;
   editor: CoreEditor | undefined;
   preview: { monomer: undefined | MonomerItemType; style: string };
-  snakeMode: boolean;
 }
 
 const initialState: EditorState = {
@@ -33,7 +32,6 @@ const initialState: EditorState = {
   activeTool: 'select',
   editor: undefined,
   preview: { monomer: undefined, style: '' },
-  snakeMode: false,
 };
 
 export const editorSlice: Slice = createSlice({
@@ -51,10 +49,6 @@ export const editorSlice: Slice = createSlice({
     },
     selectTool: (state, action: PayloadAction<string>) => {
       state.activeTool = action.payload;
-    },
-
-    selectMode: (state, action: PayloadAction<boolean>) => {
-      state.snakeMode = action.payload;
     },
 
     createEditor: (
@@ -93,7 +87,6 @@ export const {
   initSuccess,
   initFailure,
   selectTool,
-  selectMode,
   createEditor,
   showPreview,
   destroyEditor,
@@ -103,8 +96,6 @@ export const selectEditorIsReady = (state: RootState) => state.editor.isReady;
 export const selectShowPreview = (state: RootState) => state.editor.preview;
 export const selectEditorActiveTool = (state: RootState) =>
   state.editor.activeTool;
-export const selectEditorBondMode = (state: RootState) =>
-  state.editor.snakeMode;
 
 export const selectEditor = (state: RootState): CoreEditor =>
   state.editor.editor;
